infrastructure/expected-precondition-failed/precondition-failed.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test that should be skipped</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function(t) {
    assert_precondition(false);
}, 'test (synchronous)');
</script>
